requirements.txt
Django~=4.2.13
djangorestframework==3.15.2
Markdown==3.6
django-filter==24.2
django-simple-history==3.7.0